requirements.txt
Flask==3.1.1
google-cloud-bigquery==3.38.0
gunicorn==23.0.0

